Question box for question overview page
Here is a question box for the question overviews. The arrows change colors as the user hovers them for voting, there are the number of replies and the username of the one who asked the question to the left of it. Unfortunately, I have not found anything to fix the format of the title if the user shrinks the window yet. If anything is missing, let me know and I will add it.